+++ lib/self.dart
@@ -1,1 +1,0 @@
-export 'presentation/screens/slider_captcha.dart';