+++ lib/src/pizzule_path.dart
@@ -1,32 +1,4 @@
-import 'dart:typed_data';
-import 'dart:ui';
-
-import 'package:flutter/material.dart' hide Image;
-
-
-class CreateCaptChar {
-
-  static Future<Uint8List> create(
-      int width, int height, double offsetX, double offsetY) async {
-    PictureRecorder pictureRecorder = PictureRecorder();
-    Canvas canvas = Canvas(pictureRecorder);
-    final Paint paint = Paint()
-      ..color = Colors.blue
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 3.0;
-
-    canvas.drawPath(
-        getPiecePathCustom(Size(width.toDouble(), height.toDouble()), offsetX,
-            offsetY, width.toDouble()),
-        paint);
-
-    // canvas.drawImage(image, offset, paint);
-    Picture picture = pictureRecorder.endRecording();
-    final img = await picture.toImage(300, 300);
-    final pngBytes = await img.toByteData(format:ImageByteFormat.png);
-    return Uint8List.view(pngBytes!.buffer);
-  }
-}
+import 'package:flutter/material.dart';
 
 class PuzzlePiecePainter extends CustomPainter {
   PuzzlePiecePainter(this.width, this.height, this.offsetX, this.offsetY,
@@ -44,19 +16,12 @@ class PuzzlePiecePainter extends CustomPainter {
 
   @override
   Future<void> paint(Canvas canvas, Size size) async {
-    // PictureRecorder pictureRecorder = PictureRecorder();
-    // Canvas canvas = Canvas(pictureRecorder);
-
     final Paint paint = Paint()
       ..color = Colors.blue
       ..style = paintingStyle
       ..strokeWidth = 3.0;
 
     canvas.drawPath(getPiecePathCustom(size, offsetX, offsetY, width), paint);
-    // Picture picture = pictureRecorder.endRecording();
-    //
-    // // Image image =
-    // imageGlobal = (await picture.toImage(width.toInt(), height.toInt())) as Image;
   }
 
   @override
@@ -70,18 +35,20 @@ Path getPiecePathCustom(
   // offsetX = offsetX -sizePart;
 
   Path path = Path();
+
   path.moveTo(offsetX, offsetY);
 
   // top bump
   path.lineTo(offsetX + sizePart / 3, offsetY);
 
   path.cubicTo(
-      offsetX + sizePart / 6,
-      offsetY - bumpSize,
-      offsetX + sizePart / 6 * 5,
-      offsetY - bumpSize,
-      offsetX + sizePart / 3 * 2,
-      offsetY);
+    offsetX + sizePart / 6,
+    offsetY + bumpSize,
+    offsetX + sizePart / 6 * 5,
+    offsetY + bumpSize,
+    offsetX + sizePart / 3 * 2,
+    offsetY,
+  );
 
   path.lineTo(offsetX + sizePart, offsetY);
 
@@ -89,9 +56,9 @@ Path getPiecePathCustom(
   path.lineTo(offsetX + sizePart, offsetY + sizePart / 3);
 
   path.cubicTo(
-      offsetX + sizePart - bumpSize,
+      offsetX + sizePart + bumpSize,
       offsetY + sizePart / 6,
-      offsetX + sizePart - bumpSize,
+      offsetX + sizePart + bumpSize,
       offsetY + sizePart / 6 * 5,
       offsetX + sizePart,
       offsetY + sizePart / 3 * 2);
@@ -106,9 +73,9 @@ Path getPiecePathCustom(
   path.lineTo(offsetX, offsetY + sizePart / 3 * 2);
 
   path.cubicTo(
-      offsetX - bumpSize,
+      offsetX + bumpSize,
       offsetY + sizePart / 6 * 5,
-      offsetX - bumpSize,
+      offsetX + bumpSize,
       offsetY + sizePart / 6,
       offsetX,
       offsetY + sizePart / 3);