+++ lib/utils/ticker.dart
@@ -1,7 +1,0 @@
-class Ticker {
-  const Ticker();
-  Stream<int> tick({required int ticks}) {
-    return Stream.periodic(const Duration(seconds: 1), (x) => ticks - x - 1)
-        .take(ticks);
-  }
-}