+++ CHANGELOG.md
@@ -18,3 +18,23 @@
   
 ##0.07:
   TODO: Update title style for sliderBar
+  
+## 0.08
+  TODO :
+     -Remove dependencies on other libraries
+     - Fix issue: #5
+
+## 0.1.0
+TODO :
+- Improve performance
+- Add color Options
+## 0.1.1
+TODO
+- Fix bug Puzzle block position out of picture size
+## 0.1.2
+TODO
+- Fix some bug
+## 0.1.3
+TODO
+- update lint rule
+